fix scopers & tests
