--- app/overrides/decidim/consultations/question/_vote_modal/order_by_weight.html.erb.deface
@@ -1,3 +1,0 @@
-<!--  replace 'erb[silent]:contains("question.responses.each do |response|")' -->
-
-<% question.responses.sort_by(&:weight).each do |response| %>
