+++ app/overrides/decidim/consultations/questions/_results/group_responses.html.erb.deface
