+++ app/overrides/decidim/consultations/questions/show/remove_read_more.html.erb.deface
