+++ config/initializers/decidim_overrides.rb
@@ -1,5 +1,26 @@
 # frozen_string_literal: true
 
+# extra menus defined in secrets.yml
+Decidim.menu :menu do |menu|
+  if Rails.application.secrets.menu[current_organization.host.to_sym].respond_to? :each
+    Rails.application.secrets.menu[current_organization.host.to_sym].each do |item|
+      options = {}
+      options[:position] = item[:position].to_i if item[:position]
+      options[:active] = item[:active].to_sym if item[:active]
+      options[:icon_name] = item[:icon_name].to_s if item[:icon_name]
+      if item[:if_membership_in]
+        assembly = Decidim::Assembly.find_by(slug: item[:if_membership_in])
+        options[:if] = assembly ? assembly.participatory_space_private_users.where(decidim_user_id: current_user).any? : false
+      end
+      menu.add_item item[:key], I18n.t(item[:key]), item[:url], options
+    end
+  end
+end
+
+# this middleware will detect by the URL if all calls to Assembly need to skip (or include) certain types
+Rails.configuration.middleware.use AssembliesScoper
+Rails.configuration.middleware.use ParticipatoryProcessesScoper
+
 Rails.application.config.to_prepare do
   # participatory spaces private users
   Decidim::Admin::ParticipatorySpacePrivateUserForm.include(SomEnergia::Admin::ParticipatorySpacePrivateUserFormOverride)
@@ -25,4 +46,40 @@ Rails.application.config.to_prepare do
   Decidim::Consultations::Admin::UpdateResponse.include(SomEnergia::Consultations::Admin::UpdateResponseOverride)
   Decidim::Consultations::Admin::QuestionConfigurationForm.include(SomEnergia::Consultations::Admin::QuestionConfigurationFormOverride)
   Decidim::Consultations::Admin::ResponseForm.include(SomEnergia::Consultations::Admin::ResponseFormOverride)
+
+  # Assemblies alternative types
+  Decidim::Assembly.include(SomEnergia::AssemblyOverride)
+  # Participatory Processes alternative menu
+  Decidim::ParticipatoryProcess.include(SomEnergia::ParticipatoryProcessOverride)
+
+  # Creates a new menu next to Assemblies for every type configured
+  AssembliesScoper.alternative_assembly_types.each do |item|
+    Decidim.menu :menu do |menu|
+      menu.add_item item[:key],
+                    I18n.t(item[:key], scope: "decidim.assemblies.alternative_assembly_types"),
+                    Rails.application.routes.url_helpers.send("#{item[:key]}_path"),
+                    position: item[:position_in_menu],
+                    if: Decidim::Assembly.unscoped.where(organization: current_organization, assembly_type: item[:assembly_type_ids]).published.any?,
+                    active: :inclusive
+    end
+  end
+
+  # Creates a new menu next to Processes for every type configured
+  ParticipatoryProcessesScoper.scoped_participatory_process_slug_prefixes.each do |item|
+    Decidim.menu :menu do |menu|
+      menu.add_item item[:key],
+                    I18n.t(item[:key], scope: "decidim.participatory_processes.scoped_participatory_process_slug_prefixes"),
+                    Rails.application.routes.url_helpers.send("#{item[:key]}_path"),
+                    position: item[:position_in_menu],
+                    if: (
+                      Decidim::ParticipatoryProcess
+                      .unscoped
+                      .where(organization: current_organization)
+                      .where("slug LIKE ANY (ARRAY[?])", item[:slug_prefixes].map { |slug_prefix| "#{slug_prefix}%" })
+                      .published
+                      .any?
+                    ),
+                    active: :inclusive
+    end
+  end
 end
