+++ config/secrets.yml
@@ -114,16 +114,6 @@ storage_default: &storage_default
 default: &default
   decidim:
     <<: *decidim_default
-  alternative_assembly_types:
-    - 
-      key: local_groups # used to search a I18n key and a route path
-      position_in_menu: 2.6
-      assembly_type_ids: [5]
-  scoped_participatory_process_slug_prefixes:
-    -
-      key: general_assemblies # used to search a I18n key and a route path
-      position_in_menu: 2.1
-      slug_prefixes: ["SomAG"]
   omniauth:
     facebook:
       # It must be a boolean. Remember ENV variables doesn't support booleans.
@@ -167,9 +157,9 @@ default: &default
   menu:
     localhost:
       - key: somenergia.menu.team_et
-        url: /assemblies/nulla-aut
+        url: /assemblies/et
         position: 10
-        if_membership_in: nulla-aut
+        if_membership_in: et
     prod.participa.somenergia.coop:
       - key: somenergia.menu.team_et
         url: /assemblies/et
@@ -185,6 +175,16 @@ default: &default
         url: /assemblies/et
         position: 10
         if_membership_in: et
+  alternative_assembly_types:
+    - 
+      key: local_groups # used to search a I18n key and a route path
+      position_in_menu: 2.6
+      assembly_type_ids: [5]
+  scoped_participatory_process_slug_prefixes:
+    -
+      key: general_assemblies # used to search a I18n key and a route path
+      position_in_menu: 2.1
+      slug_prefixes: ["SomAG"]
 
 development:
   <<: *default
